- added chmod wp  - added are you sure to drop.sh  - changed pass for go2.sh to fit

+++ drop.sh
@@ -25,6 +25,13 @@ if [ $# -eq 0 ]
 	exit 0;
 fi
 
-echo "Exporting database $1"
-mysql -u $mysql_user -p"${mysql_password}" -e "Drop database ${1};" 
-echo "Database dropped"
+
+read -p "Are you sure? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "dropping database $1"
+    mysql -u $mysql_user -p"${mysql_password}" -e "Drop database ${1};"
+    echo "Database dropped"
+fi
+
